refactor: complete crate boundary cleanup and semantic alignment
