+++ crates/temper-cli/src/serve/mod.rs
@@ -299,6 +299,10 @@ fn spawn_optimization_loop(state: &PlatformState) {
     });
 }
 
+fn is_ephemeral_metadata_error(err: &str) -> bool {
+    err.contains("explicit ephemeral mode")
+}
+
 /// Spawn background verification tasks for each entity in the specs directory.
 ///
 /// For each entity, runs the verification cascade in a blocking task and
@@ -351,10 +355,14 @@ async fn spawn_background_verification(state: &PlatformState, specs_dir: &str, t
                 .persist_spec_verification(&tenant, &entity, "running", None)
                 .await
             {
-                eprintln!(
-                    "Warning: failed to persist running verification status for {tenant}/{entity}: {e}"
-                );
-                return;
+                if is_ephemeral_metadata_error(&e) {
+                    eprintln!("Info: {tenant}/{entity} verification status is in-memory only: {e}");
+                } else {
+                    eprintln!(
+                        "Warning: failed to persist running verification status for {tenant}/{entity}: {e}"
+                    );
+                    return;
+                }
             }
             {
                 let tenant_id = TenantId::new(&tenant);
@@ -460,10 +468,16 @@ async fn spawn_background_verification(state: &PlatformState, specs_dir: &str, t
                         )
                         .await
                     {
-                        eprintln!(
-                            "Warning: failed to persist final verification status for {tenant}/{entity}: {e}"
-                        );
-                        return;
+                        if is_ephemeral_metadata_error(&e) {
+                            eprintln!(
+                                "Info: {tenant}/{entity} final verification status is in-memory only: {e}"
+                            );
+                        } else {
+                            eprintln!(
+                                "Warning: failed to persist final verification status for {tenant}/{entity}: {e}"
+                            );
+                            return;
+                        }
                     }
                     {
                         let tenant_id = TenantId::new(&tenant);
@@ -522,10 +536,16 @@ async fn spawn_background_verification(state: &PlatformState, specs_dir: &str, t
                         )
                         .await
                     {
-                        eprintln!(
-                            "Warning: failed to persist failed verification status for {tenant}/{entity_clone}: {persist_err}"
-                        );
-                        return;
+                        if is_ephemeral_metadata_error(&persist_err) {
+                            eprintln!(
+                                "Info: {tenant}/{entity_clone} failed verification status is in-memory only: {persist_err}"
+                            );
+                        } else {
+                            eprintln!(
+                                "Warning: failed to persist failed verification status for {tenant}/{entity_clone}: {persist_err}"
+                            );
+                            return;
+                        }
                     }
                     {
                         let tenant_id = TenantId::new(&tenant);
