+++ crates/temper-platform/tests/common/mod.rs
@@ -1,0 +1,6 @@
+#![allow(dead_code)]
+
+pub mod dst;
+pub mod http;
+pub mod platform;
+pub mod specs;
