+++ crates/temper-server/src/odata/write.rs
@@ -1,0 +1,484 @@
+//! OData write handlers (`POST`, `PATCH`, `PUT`, `DELETE`).
+
+use axum::extract::State;
+use axum::http::{HeaderMap, StatusCode};
+use axum::response::IntoResponse;
+use temper_odata::path::{ODataPath, parse_path};
+use temper_runtime::tenant::TenantId;
+
+use super::bindings::dispatch_bound_action;
+use super::common::{
+    constraint_violation_response, extract_key, extract_tenant, resolve_entity_type,
+    verification_gate_response,
+};
+use super::response::annotate_entity;
+use crate::constraint_engine::{
+    post_write_invariant_checks, pre_delete_relation_checks, pre_upsert_relation_checks,
+};
+use crate::response::{ODataResponse, odata_error};
+use crate::state::ServerState;
+
+fn parse_odata_path_or_400(path: &str) -> Result<ODataPath, axum::response::Response> {
+    parse_path(&format!("/{path}")).map_err(|e| {
+        odata_error(StatusCode::BAD_REQUEST, "InvalidPath", &e.to_string()).into_response()
+    })
+}
+
+fn parse_json_body_or_400(
+    body: &axum::body::Bytes,
+) -> Result<serde_json::Value, axum::response::Response> {
+    serde_json::from_slice(body).map_err(|e| {
+        odata_error(
+            StatusCode::BAD_REQUEST,
+            "InvalidBody",
+            &format!("Invalid JSON body: {e}"),
+        )
+        .into_response()
+    })
+}
+
+fn resolve_entity_type_or_404(
+    state: &ServerState,
+    tenant: &TenantId,
+    set_name: &str,
+) -> Result<String, axum::response::Response> {
+    resolve_entity_type(state, tenant, set_name).ok_or_else(|| {
+        odata_error(
+            StatusCode::NOT_FOUND,
+            "EntitySetNotFound",
+            &format!("Entity set '{set_name}' not found"),
+        )
+        .into_response()
+    })
+}
+
+fn check_verification_gate_or_423(
+    state: &ServerState,
+    tenant: &TenantId,
+    entity_type: &str,
+) -> Result<(), axum::response::Response> {
+    state
+        .check_verification_gate(tenant, entity_type)
+        .map_err(verification_gate_response)
+}
+
+fn ensure_entity_exists_or_404(
+    state: &ServerState,
+    tenant: &TenantId,
+    entity_type: &str,
+    set_name: &str,
+    key: &str,
+) -> Result<(), axum::response::Response> {
+    if state.entity_exists(tenant, entity_type, key) {
+        Ok(())
+    } else {
+        Err(odata_error(
+            StatusCode::NOT_FOUND,
+            "ResourceNotFound",
+            &format!("Entity '{set_name}' with key '{key}' not found"),
+        )
+        .into_response())
+    }
+}
+
+/// Handle POST requests — entity creation and bound actions.
+pub async fn handle_odata_post(
+    State(state): State<ServerState>,
+    headers: HeaderMap,
+    axum::extract::Path(path): axum::extract::Path<String>,
+    body: axum::body::Bytes,
+) -> impl IntoResponse {
+    let tenant = extract_tenant(&headers, &state);
+    let odata_path = match parse_odata_path_or_400(&path) {
+        Ok(p) => p,
+        Err(resp) => return resp,
+    };
+
+    match odata_path {
+        ODataPath::EntitySet(name) => {
+            let entity_type = match resolve_entity_type_or_404(&state, &tenant, &name) {
+                Ok(t) => t,
+                Err(resp) => return resp,
+            };
+            if let Err(resp) = check_verification_gate_or_423(&state, &tenant, &entity_type) {
+                return resp;
+            }
+
+            let body_json = match parse_json_body_or_400(&body) {
+                Ok(v) => v,
+                Err(resp) => return resp,
+            };
+
+            let entity_id = body_json
+                .get("id")
+                .and_then(|v| v.as_str())
+                .map(String::from)
+                .unwrap_or_else(|| temper_runtime::scheduler::sim_uuid().to_string());
+
+            let initial_fields = body_json.clone();
+            if let Err(v) = pre_upsert_relation_checks(
+                &state,
+                &tenant,
+                &entity_type,
+                &entity_id,
+                "create",
+                &initial_fields,
+            )
+            .await
+            {
+                return constraint_violation_response(v);
+            }
+            if let Err(v) = post_write_invariant_checks(
+                &state,
+                &tenant,
+                &entity_type,
+                &entity_id,
+                "Create",
+                &initial_fields,
+                "create",
+            )
+            .await
+            {
+                return constraint_violation_response(v);
+            }
+
+            match state
+                .get_or_create_tenant_entity(&tenant, &entity_type, &entity_id, initial_fields)
+                .await
+            {
+                Ok(response) => {
+                    let body = annotate_entity(
+                        serde_json::to_value(&response.state).unwrap_or_default(),
+                        format!("$metadata#{name}/$entity"),
+                        Some(format!("{name}('{entity_id}')")),
+                    );
+                    ODataResponse {
+                        status: StatusCode::CREATED,
+                        body,
+                    }
+                    .into_response()
+                }
+                Err(e) => odata_error(StatusCode::INTERNAL_SERVER_ERROR, "CreateError", &e)
+                    .into_response(),
+            }
+        }
+
+        ODataPath::BoundAction { parent, action } => {
+            let body_json: serde_json::Value = serde_json::from_slice(&body).unwrap_or_default();
+
+            let (set_name, key_str) = match *parent {
+                ODataPath::Entity(ref set, ref key) => (set.clone(), extract_key(key)),
+                _ => {
+                    return odata_error(
+                        StatusCode::BAD_REQUEST,
+                        "InvalidPath",
+                        "Action must be bound to an entity",
+                    )
+                    .into_response();
+                }
+            };
+
+            let entity_type = match resolve_entity_type_or_404(&state, &tenant, &set_name) {
+                Ok(t) => t,
+                Err(resp) => return resp,
+            };
+
+            if let Err(resp) = check_verification_gate_or_423(&state, &tenant, &entity_type) {
+                return resp;
+            }
+            dispatch_bound_action(
+                &state,
+                &tenant,
+                &set_name,
+                &entity_type,
+                &key_str,
+                &action,
+                body_json,
+            )
+            .await
+        }
+
+        _ => odata_error(
+            StatusCode::METHOD_NOT_ALLOWED,
+            "MethodNotAllowed",
+            "POST not supported for this path",
+        )
+        .into_response(),
+    }
+}
+
+/// Handle PATCH requests — partial entity update.
+pub async fn handle_odata_patch(
+    State(state): State<ServerState>,
+    headers: HeaderMap,
+    axum::extract::Path(path): axum::extract::Path<String>,
+    body: axum::body::Bytes,
+) -> impl IntoResponse {
+    let tenant = extract_tenant(&headers, &state);
+    let odata_path = match parse_odata_path_or_400(&path) {
+        Ok(p) => p,
+        Err(resp) => return resp,
+    };
+
+    match odata_path {
+        ODataPath::Entity(set_name, key) => {
+            let entity_type = match resolve_entity_type_or_404(&state, &tenant, &set_name) {
+                Ok(t) => t,
+                Err(resp) => return resp,
+            };
+            let key_str = extract_key(&key);
+
+            if let Err(resp) = check_verification_gate_or_423(&state, &tenant, &entity_type) {
+                return resp;
+            }
+            if let Err(resp) =
+                ensure_entity_exists_or_404(&state, &tenant, &entity_type, &set_name, &key_str)
+            {
+                return resp;
+            }
+
+            let body_json = match parse_json_body_or_400(&body) {
+                Ok(v) => v,
+                Err(resp) => return resp,
+            };
+            let current_state = match state
+                .get_tenant_entity_state(&tenant, &entity_type, &key_str)
+                .await
+            {
+                Ok(v) => v,
+                Err(e) => {
+                    return odata_error(StatusCode::INTERNAL_SERVER_ERROR, "ReadError", &e)
+                        .into_response();
+                }
+            };
+
+            let mut prospective_fields = current_state.state.fields.clone();
+            if let (Some(dst), Some(src)) =
+                (prospective_fields.as_object_mut(), body_json.as_object())
+            {
+                for (k, v) in src {
+                    dst.insert(k.clone(), v.clone());
+                }
+            } else {
+                prospective_fields = body_json.clone();
+            }
+
+            if let Err(v) = pre_upsert_relation_checks(
+                &state,
+                &tenant,
+                &entity_type,
+                &key_str,
+                "patch",
+                &prospective_fields,
+            )
+            .await
+            {
+                return constraint_violation_response(v);
+            }
+            if let Err(v) = post_write_invariant_checks(
+                &state,
+                &tenant,
+                &entity_type,
+                &key_str,
+                "Patch",
+                &prospective_fields,
+                "patch",
+            )
+            .await
+            {
+                return constraint_violation_response(v);
+            }
+
+            match state
+                .update_tenant_entity_fields(&tenant, &entity_type, &key_str, body_json, false)
+                .await
+            {
+                Ok(response) => {
+                    let body = annotate_entity(
+                        serde_json::to_value(&response.state).unwrap_or_default(),
+                        format!("$metadata#{set_name}/$entity"),
+                        Some(format!("{set_name}('{key_str}')")),
+                    );
+                    ODataResponse {
+                        status: StatusCode::OK,
+                        body,
+                    }
+                    .into_response()
+                }
+                Err(e) => odata_error(StatusCode::INTERNAL_SERVER_ERROR, "UpdateError", &e)
+                    .into_response(),
+            }
+        }
+        _ => odata_error(
+            StatusCode::METHOD_NOT_ALLOWED,
+            "MethodNotAllowed",
+            "PATCH only supported on entity instances",
+        )
+        .into_response(),
+    }
+}
+
+/// Handle PUT requests — full entity replacement.
+pub async fn handle_odata_put(
+    State(state): State<ServerState>,
+    headers: HeaderMap,
+    axum::extract::Path(path): axum::extract::Path<String>,
+    body: axum::body::Bytes,
+) -> impl IntoResponse {
+    let tenant = extract_tenant(&headers, &state);
+    let odata_path = match parse_odata_path_or_400(&path) {
+        Ok(p) => p,
+        Err(resp) => return resp,
+    };
+
+    match odata_path {
+        ODataPath::Entity(set_name, key) => {
+            let entity_type = match resolve_entity_type_or_404(&state, &tenant, &set_name) {
+                Ok(t) => t,
+                Err(resp) => return resp,
+            };
+            let key_str = extract_key(&key);
+
+            if let Err(resp) = check_verification_gate_or_423(&state, &tenant, &entity_type) {
+                return resp;
+            }
+            if let Err(resp) =
+                ensure_entity_exists_or_404(&state, &tenant, &entity_type, &set_name, &key_str)
+            {
+                return resp;
+            }
+
+            let body_json = match parse_json_body_or_400(&body) {
+                Ok(v) => v,
+                Err(resp) => return resp,
+            };
+
+            if let Err(v) = pre_upsert_relation_checks(
+                &state,
+                &tenant,
+                &entity_type,
+                &key_str,
+                "put",
+                &body_json,
+            )
+            .await
+            {
+                return constraint_violation_response(v);
+            }
+            if let Err(v) = post_write_invariant_checks(
+                &state,
+                &tenant,
+                &entity_type,
+                &key_str,
+                "Put",
+                &body_json,
+                "put",
+            )
+            .await
+            {
+                return constraint_violation_response(v);
+            }
+
+            match state
+                .update_tenant_entity_fields(&tenant, &entity_type, &key_str, body_json, true)
+                .await
+            {
+                Ok(response) => {
+                    let body = annotate_entity(
+                        serde_json::to_value(&response.state).unwrap_or_default(),
+                        format!("$metadata#{set_name}/$entity"),
+                        Some(format!("{set_name}('{key_str}')")),
+                    );
+                    ODataResponse {
+                        status: StatusCode::OK,
+                        body,
+                    }
+                    .into_response()
+                }
+                Err(e) => odata_error(StatusCode::INTERNAL_SERVER_ERROR, "UpdateError", &e)
+                    .into_response(),
+            }
+        }
+        _ => odata_error(
+            StatusCode::METHOD_NOT_ALLOWED,
+            "MethodNotAllowed",
+            "PUT only supported on entity instances",
+        )
+        .into_response(),
+    }
+}
+
+/// Handle DELETE requests — entity deletion.
+pub async fn handle_odata_delete(
+    State(state): State<ServerState>,
+    headers: HeaderMap,
+    axum::extract::Path(path): axum::extract::Path<String>,
+) -> impl IntoResponse {
+    let tenant = extract_tenant(&headers, &state);
+    let odata_path = match parse_odata_path_or_400(&path) {
+        Ok(p) => p,
+        Err(resp) => return resp,
+    };
+
+    match odata_path {
+        ODataPath::Entity(set_name, key) => {
+            let entity_type = match resolve_entity_type_or_404(&state, &tenant, &set_name) {
+                Ok(t) => t,
+                Err(resp) => return resp,
+            };
+            let key_str = extract_key(&key);
+
+            if let Err(resp) = check_verification_gate_or_423(&state, &tenant, &entity_type) {
+                return resp;
+            }
+            if let Err(resp) =
+                ensure_entity_exists_or_404(&state, &tenant, &entity_type, &set_name, &key_str)
+            {
+                return resp;
+            }
+            if let Err(v) =
+                pre_delete_relation_checks(&state, &tenant, &entity_type, &key_str, "delete").await
+            {
+                return constraint_violation_response(v);
+            }
+            let current_state = match state
+                .get_tenant_entity_state(&tenant, &entity_type, &key_str)
+                .await
+            {
+                Ok(v) => v,
+                Err(e) => {
+                    return odata_error(StatusCode::INTERNAL_SERVER_ERROR, "ReadError", &e)
+                        .into_response();
+                }
+            };
+            if let Err(v) = post_write_invariant_checks(
+                &state,
+                &tenant,
+                &entity_type,
+                &key_str,
+                "Delete",
+                &current_state.state.fields,
+                "delete",
+            )
+            .await
+            {
+                return constraint_violation_response(v);
+            }
+
+            match state
+                .delete_tenant_entity(&tenant, &entity_type, &key_str)
+                .await
+            {
+                Ok(_) => (StatusCode::NO_CONTENT, "").into_response(),
+                Err(e) => odata_error(StatusCode::INTERNAL_SERVER_ERROR, "DeleteError", &e)
+                    .into_response(),
+            }
+        }
+        _ => odata_error(
+            StatusCode::METHOD_NOT_ALLOWED,
+            "MethodNotAllowed",
+            "DELETE only supported on entity instances",
+        )
+        .into_response(),
+    }
+}
